Paddle reworked
